lib/com_pac/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "com_pac"
version = "0.1.0"
description = "A configurable communication library for TCP, UDP, UART, and Ethernet."
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = { file = "LICENSE" }
dependencies = ["pyserial"]
